AWS: allow dots ('.') in bucket name.
Previously, HTTPS certificate validation would fail if dots were
in the bucket name because the bucketURL would then contain multiple
levels of subdomains (e.g. some.depth.of.domains.s3.amazonaws.com).

Now, the bucket name is not included in the domain name, and the HTTPS
certificate is valid even when dots are in the bucket name.